+++ src/NuGet.Services.KeyVault/NuGet.Services.KeyVault.csproj
@@ -52,6 +52,7 @@
     <Compile Include="ISecretReader.cs" />
     <Compile Include="KeyVaultConfiguration.cs" />
     <Compile Include="KeyVaultReader.cs" />
+    <Compile Include="KeyVaultSecret.cs" />
     <Compile Include="SecretInjector.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.*.cs" />
